Added basic idf.py v5 support

+++ src/i2cdev.c
@@ -197,7 +197,7 @@ static esp_err_t i2c_setup_port(const i2c_dev_t *dev)
     esp_err_t res;
     if (!cfg_equal(&dev->cfg, &states[dev->port].config))
     {
-        ESP_LOGD(TAG, "Reconfiguring I2C driver on port %d", dev->port);
+        ESP_LOGD(TAG, "Reconfiguring I2C driver on port %d", (int)dev->port);
         i2c_config_t temp;
         memcpy(&temp, &dev->cfg, sizeof(i2c_config_t));
         temp.mode = I2C_MODE_MASTER;
@@ -222,7 +222,7 @@ static esp_err_t i2c_setup_port(const i2c_dev_t *dev)
         states[dev->port].installed = true;
 
         memcpy(&states[dev->port].config, &temp, sizeof(i2c_config_t));
-        ESP_LOGD(TAG, "I2C driver successfully reconfigured on port %d", dev->port);
+        ESP_LOGD(TAG, "I2C driver successfully reconfigured on port %d", (int)dev->port);
     }
 #if HELPER_TARGET_IS_ESP32
     int t;
@@ -232,7 +232,7 @@ static esp_err_t i2c_setup_port(const i2c_dev_t *dev)
     uint32_t ticks = dev->timeout_ticks ? dev->timeout_ticks : I2CDEV_MAX_STRETCH_TIME;
     if ((ticks != t) && (res = i2c_set_timeout(dev->port, ticks)) != ESP_OK)
         return res;
-    ESP_LOGD(TAG, "Timeout: ticks = %d (%d usec) on port %d", dev->timeout_ticks, dev->timeout_ticks / 80, dev->port);
+    ESP_LOGD(TAG, "Timeout: ticks = %d (%d usec) on port %d", (int)dev->timeout_ticks, (int)(dev->timeout_ticks / 80), (int)dev->port);
 #endif
 
     return ESP_OK;
